Promenjen ispis, mark_indentation prebačen u util.py
