+++ ast_nodes.py
@@ -1,5 +1,3 @@
-# ast_nodes.py
-
 from dataclasses import dataclass, field
 from typing import List, Optional, Union
